requirements.txt
Django==5.1.4
sqlparse==0.5.0
Pillow==10.3.0
flower==2.0.1
celery==5.4.0
stripe==9.3.0
python-decouple==3.8
WeasyPrint==61.2
redis==5.0.4
django-rosetta==0.10.0
django-parler==2.3
django-localflavor==4.0
django-colorfield
django-redis
psycopg2-binary
dj-database-url
whitenoise==6.2.0
Brotli==1.0.9
gunicorn
uvicorn
django_crispy_forms
django_widget_tweaks
